refactor: constants, types 변경
