--- src/constants/numbers.ts
@@ -1,2 +1,0 @@
-export const MAX_TITLE_LENGTH = 80;
-export const MAX_DESCRIPTION_LENGTH = 100;
