+++ src/types/article.d.ts
@@ -1,0 +1,15 @@
+interface ArticleSource {
+  id: string | null;
+  name: string;
+}
+
+export interface Article {
+  source: ArticleSource;
+  author: string | null;
+  title: string;
+  description: string | null;
+  url: string;
+  urlToImage: string | null;
+  publishedAt: string;
+  content: string | null;
+}
