+++ src/types/data.d.ts
@@ -1,0 +1,8 @@
+import { Article } from './article';
+import { Category } from './category';
+
+interface NewsDataStore {
+  newsData: Article[];
+  setNewsData: (category: Category) => Promise<void>;
+  isLoading: boolean;
+}
